requirements.txt
# Core
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic[email]>=2.6
pydantic-settings>=2.2

# Auto load .env files from root directory
python-dotenv>=1.0
httpx>=0.27
rich>=13.7
plotly>=6.5.2
# Streamlit UI
streamlit>=1.31

# Agents / graph
langgraph>=0.2.30
langchain-core>=0.3.0

# LLM providers
anthropic>=0.86.0
openai>=1.40.0

pdfplumber>=0.11.9

# Azure Document Intelligence SDK
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0

python-multipart>=0.0.21

pdf2image>=1.17.0

# PDF generation (test paystub generation)
reportlab>=4.0

# Stock market data
yfinance>=0.2.40

# Background scheduler (stock alert polling)
apscheduler>=3.10,<4.0

# Testing
pytest>=8.0